requirements.txt
streamlit
anthropic
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-docx
rapidfuzz
pdfplumber
